+++ src/main/java/com/moyo/backend/domain/temporary_batch/GithubCommitStats.java
@@ -1,4 +1,4 @@
-package com.moyo.backend.domain.github_ranking.temporary_batch;
+package com.moyo.backend.domain.temporary_batch;
 
 public record GithubCommitStats(
 	WeekStats weekStats,
