+++ src/main/java/com/moyo/backend/domain/temporary_batch/RankingBatchReader.java
@@ -1,4 +1,4 @@
-package com.moyo.backend.domain.github_ranking.temporary_batch;
+package com.moyo.backend.domain.temporary_batch;
 
 import static com.moyo.backend.common.constant.MoyoConstants.GITHUB_MIN_REQUEST_THRESHOLD;
 
@@ -17,6 +17,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.moyo.backend.common.exception.github_follow.GithubRateLimitExceedException;
+import com.moyo.backend.domain.temporary_batch.data_access.ContributorStats;
+import com.moyo.backend.domain.temporary_batch.data_access.GithubRankingHttpClient;
 
 @Slf4j
 @Component
