+++ src/main/java/com/moyo/backend/domain/temporary_batch/data_access/GithubRankingHttpClient.java
@@ -1,8 +1,12 @@
-package com.moyo.backend.domain.github_ranking.temporary_batch;
+package com.moyo.backend.domain.temporary_batch.data_access;
 
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
+
+import com.moyo.backend.domain.temporary_batch.GithubContributorDetails;
+import com.moyo.backend.domain.temporary_batch.GithubRepoDetails;
+import com.moyo.backend.domain.temporary_batch.RankingPreflight;
 
 public interface GithubRankingHttpClient {
